Update files : Change bootstrap style

+++ var/cache/dev/twig/99/991743f3998f7483b431c3ae66e3e9c955ecbe3a4856d8ac382bd577f050396b.php
@@ -96,46 +96,57 @@ class __TwigTemplate_8f107b59316735fdbb3cc2e77c88b6689666986d84e24a29515ad95234e
 
     <div class=\"container\">
         <div class=\"jumbotron\">
-            <h1>Liste d'utilisateurs</h1>
             ";
-        // line 18
-        $context['_parent'] = $context;
-        $context['_seq'] = twig_ensure_traversable((isset($context["users"]) || array_key_exists("users", $context) ? $context["users"] : (function () { throw new RuntimeError('Variable "users" does not exist.', 18, $this->source); })()));
-        foreach ($context['_seq'] as $context["_key"] => $context["user"]) {
+        // line 17
+        if ((isset($context["users"]) || array_key_exists("users", $context) ? $context["users"] : (function () { throw new RuntimeError('Variable "users" does not exist.', 17, $this->source); })())) {
+            // line 18
+            echo "            <h1>Liste d'utilisateurs</h1>
+            ";
             // line 19
-            echo "                <div class=\"card-body bg-light m-5\">
+            $context['_parent'] = $context;
+            $context['_seq'] = twig_ensure_traversable((isset($context["users"]) || array_key_exists("users", $context) ? $context["users"] : (function () { throw new RuntimeError('Variable "users" does not exist.', 19, $this->source); })()));
+            foreach ($context['_seq'] as $context["_key"] => $context["user"]) {
+                // line 20
+                echo "                <div class=\"card-body bg-light m-5\">
                     <h5 class=\"card-title\">";
-            // line 20
-            echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "name", [], "any", false, false, false, 20), "html", null, true);
-            echo " ";
-            echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "firstname", [], "any", false, false, false, 20), "html", null, true);
-            echo "</h5>
+                // line 21
+                echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "name", [], "any", false, false, false, 21), "html", null, true);
+                echo " ";
+                echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "firstname", [], "any", false, false, false, 21), "html", null, true);
+                echo "</h5>
                     <p class=\"card-text\">";
-            // line 21
-            echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "mail", [], "any", false, false, false, 21), "html", null, true);
-            echo "</p>
+                // line 22
+                echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "mail", [], "any", false, false, false, 22), "html", null, true);
+                echo "</p>
                     <p class='card-text'>Crée le ";
-            // line 22
-            echo twig_escape_filter($this->env, twig_date_format_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "createdAt", [], "any", false, false, false, 22), "d/m/Y"), "html", null, true);
-            echo "</p>
+                // line 23
+                echo twig_escape_filter($this->env, twig_date_format_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "createdAt", [], "any", false, false, false, 23), "d/m/Y"), "html", null, true);
+                echo "</p>
                     <a href=\"user/";
-            // line 23
-            echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "id", [], "any", false, false, false, 23), "html", null, true);
-            echo "\" class=\"btn btn-primary\">Voir plus</a>
+                // line 24
+                echo twig_escape_filter($this->env, twig_get_attribute($this->env, $this->source, $context["user"], "id", [], "any", false, false, false, 24), "html", null, true);
+                echo "\" class=\"btn btn-primary\">Voir plus</a>
                 </div>
             ";
+            }
+            $_parent = $context['_parent'];
+            unset($context['_seq'], $context['_iterated'], $context['_key'], $context['user'], $context['_parent'], $context['loop']);
+            $context = array_intersect_key($context, $_parent) + $_parent;
+            // line 27
+            echo "            ";
+        } else {
+            // line 28
+            echo "            <h1>Il n'y a pas d'utilisateur</h1>
+            ";
         }
-        $_parent = $context['_parent'];
-        unset($context['_seq'], $context['_iterated'], $context['_key'], $context['user'], $context['_parent'], $context['loop']);
-        $context = array_intersect_key($context, $_parent) + $_parent;
-        // line 26
+        // line 30
         echo "        </div>
 
         <div class=\"jumbotron\">
             <h2>Ajouter utilisateur</h2>
         ";
-        // line 30
-        echo         $this->env->getRuntime('Symfony\Component\Form\FormRenderer')->renderBlock((isset($context["formUsers"]) || array_key_exists("formUsers", $context) ? $context["formUsers"] : (function () { throw new RuntimeError('Variable "formUsers" does not exist.', 30, $this->source); })()), 'form');
+        // line 34
+        echo         $this->env->getRuntime('Symfony\Component\Form\FormRenderer')->renderBlock((isset($context["formUsers"]) || array_key_exists("formUsers", $context) ? $context["formUsers"] : (function () { throw new RuntimeError('Variable "formUsers" does not exist.', 34, $this->source); })()), 'form');
         echo "
         </div>
     </div>
@@ -160,7 +171,7 @@ class __TwigTemplate_8f107b59316735fdbb3cc2e77c88b6689666986d84e24a29515ad95234e
 
     public function getDebugInfo()
     {
-        return array (  138 => 30,  132 => 26,  123 => 23,  119 => 22,  115 => 21,  109 => 20,  106 => 19,  102 => 18,  88 => 6,  78 => 5,  59 => 3,  36 => 1,);
+        return array (  149 => 34,  143 => 30,  139 => 28,  136 => 27,  127 => 24,  123 => 23,  119 => 22,  113 => 21,  110 => 20,  106 => 19,  103 => 18,  101 => 17,  88 => 6,  78 => 5,  59 => 3,  36 => 1,);
     }
 
     public function getSourceContext()
@@ -181,6 +192,7 @@ class __TwigTemplate_8f107b59316735fdbb3cc2e77c88b6689666986d84e24a29515ad95234e
 
     <div class=\"container\">
         <div class=\"jumbotron\">
+            {% if users %}
             <h1>Liste d'utilisateurs</h1>
             {% for user in users %}
                 <div class=\"card-body bg-light m-5\">
@@ -190,6 +202,9 @@ class __TwigTemplate_8f107b59316735fdbb3cc2e77c88b6689666986d84e24a29515ad95234e
                     <a href=\"user/{{user.id}}\" class=\"btn btn-primary\">Voir plus</a>
                 </div>
             {% endfor %}
+            {% else %}
+            <h1>Il n'y a pas d'utilisateur</h1>
+            {% endif %}
         </div>
 
         <div class=\"jumbotron\">
